new metafile methods and long for length

+++ src/main/java/ru/gdcn/polytorrent/tracker/FileData.java
@@ -4,16 +4,16 @@ import java.util.List;
 
 public class FileData {
 
-    private final int length;
+    private final long length;
 
     private final List<String> path;
 
-    public FileData(int length, List<String> path) {
+    public FileData(long length, List<String> path) {
         this.length = length;
         this.path = path;
     }
 
-    public int getLength() {
+    public long getLength() {
         return length;
     }
